requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
h3==3.7.6
sentence-transformers==2.2.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.1
pandas==2.1.2
folium==0.14.0
